feat!: restructure artifact paths
This site is ditching the cs.jneander.dev domain and moving to
toys.jneander.dev. It is also flattening the project pages into a single
path under /artifacts. This will separate project pages from site pages,
avoiding any potential (but unlikely) name clashing between those
conceptual domains.

In reality, there will likely be name clashes between artifacts
representing variations on the same project with different scopes. There
is currently no comprehensive strategy for this, as it remains
hypothetical and absolutely has a solution of some kind. But that is
best informed by the shape of the problem when it is encountered.